refactor(cosh-ng): split session store owners
Extract the inline test suite into session/store/tests.rs.

Extract legacy directory discovery, locking, and removal into session/store/legacy.rs.

The production owner drops from 2213 to 719 lines, below the 1000-line no-growth threshold.

No behavior change: moved items keep pub(super) visibility inside the session module.
